[FIX] cd /root  [CLEAN] del chdirerr2.c  [CLEAN] des appels a exit -> ft_exit
